Explicit fix for demoweb alias
In the past the file was stashed but local changes were now lost.

I think doing this check is also a good way to autodocument how this should look like. Disadvantage is that a change in the apache.conf.in may now also result in a change in this repo.