Rewrite for 22.04 domserver
Once every 2 years there is a new ubuntu LTS and the PHP version being
used at the WF could be different from the other contests taking place
around it (because of the image being decided at a contest before). We
now use the installed version instead of a hardcoded version.

Also most copy actions are replaced with synchronize as adviced by Red
hat to make these actions faster in case of a rerun. Copy is kept for
recursive actions but we should change those also to get more speedup.

Setup the ssh user with a double loop so the root+domjudge user are in
sync.

Target all new groups for clusterssh, since SWERC we can have more
groups which we might want to manage in one go.

For each OS now set the kernel to chrootV1, if the parameter is not
implemented in the kernel this will just be ignored.

Fixed the pretask for judgedaemons to check if the service file exists
and if so to turn off the service.

Removed some redundant comments.