Use encrypt nginx/mariadb/node exporter traffic
This is not possible (yet) for php_fpm but that one also exposes the
least interesting metrics for people snooping. If we would want that we
would need to do something with nginx for this.

This is relevant for the online judge but also makes sense as teams can
always query the node endpoint on domservers.