Secure scrapes for online instances
This does share some implicit configuration from the target_all but as
we always deploy those together this should be fine.

For online instances which are directly connected to the internet we
didn't secure all metrics. This is now done for:
node (builtin)
mysql (builtin)
fpm (builtin)
nginx (Custom NGINX config)