Split for WF setups
For the world finals we both test with an online and an onprem
instance. We sometimes even manage an online server for the bigger
contests together with a server which has no internet access. This also
demonstrates to others using this how to setup something like even if we
always only use 1 group on the location.