Make sure tune_cpu gets started
In the past we could (and did) fail somewhere in the playbook which
would result in not handling this notified handler. This also meant that
we would not copy the file again on the next run, so didn't kick the
handler and also didn't hit the handler on the next run.

The disadvantage here is we might hit other handlers more often if next
roles want to hit the same handler.