Define GlitchTip installation role
In a later restructure of Ansible, the GlithTip role should become part
of the admin role, just like the Grafana roles has to be merged in.

Use the reusable Docker definitions to install Docker.
In the compose stack, use the host CA bundle for a.o. uptime monitor as
often custom certs are used on contest floors which must be trusted in
the container.